Redirect filenames/slugs with parentheses to non-parentheses equivalents
Fixes https://github.com/mdn/content/issues/13936

In the CSS substree, we have 76 articles whose filenames and slugs
contain parentheses. Those parentheses create a number of problems —
including that many or most URL auto-linkifiers don’t recognize the
parentheses as being part of the URL, and so only linkify the part
before the parentheses; one place where that happens is Twitter; see,
e.g., https://twitter.com/DctStrangelove/status/1504007889316364288

And another problem for our own maintenance is that command-line shells
don’t expect parenthesis characters in filenames — because those
characters have special meanings to shells (and to regex engines, etc.),
and so any time want to do something with those filenames or slugs for
the command line, we need to quote the filenames or slugs.